Struct golem_rpc_api::comp::GolemComp

pub struct GolemComp<'a, Inner: RpcEndpoint + ?Sized>(_);

Methods

impl<'a, Inner: RpcEndpoint + ?Sized + 'static> GolemComp<'a, Inner>

pub fn create_task_int(
&self,
task_spec: Value
) -> impl Future<Item = (Option<String>, Option<Value>), Error = Error>

RPC uri= comp.task.create

pub fn get_task(
&self,
task_id: String
) -> impl Future<Item = Option<TaskInfo>, Error = Error>

RPC uri= comp.task

pub fn get_tasks<'b>(
&'b self
) -> impl Future<Item = Vec<TaskInfo>, Error = Error> + 'static

RPC uri= comp.tasks

pub fn get_tasks_unsupported(
&self,
last_days: i32
) -> impl Future<Item = Vec<UnsupportInfo>, Error = Error>

Show statistics for unsupported tasks.

Arguments

last_days - Number of last days to compute statistics on.

Returns

Vec of UnsupportInfo. With stats for each reason.

RPC uri= comp.tasks.unsupport

pub fn abort_task(
&self,
task_id: String
) -> impl Future<Item = (), Error = Error>

Abort task with given id.

Arguments

task_id - Task id to abort.

RPC uri= comp.task.abort

pub fn delete_task(
&self,
taks_id: String
) -> impl Future<Item = (), Error = Error>

RPC uri= comp.task.delete

pub fn restart_subtask(
&self,
subtask_id: String
) -> impl Future<Item = (), Error = Error>

RPC uri= comp.task.subtask.restart

pub fn get_subtask(
&self,
subtask_id: String
) -> impl Future<Item = (Option<SubtaskInfo>, Option<String>), Error = Error>

RPC uri= comp.task.subtask

pub fn get_subtasks(
&self,
task_id: String
) -> impl Future<Item = Option<Vec<SubtaskInfo>>, Error = Error>

RPC uri= comp.task.subtasks

pub fn purge_tasks<'b>(
&'b self
) -> impl Future<Item = (), Error = Error> + 'static

RPC uri= comp.task.purge

pub fn restart_task(
&self,
task_id: String
) -> impl Future<Item = (Option<String>, Option<String>), Error = Error>

RPC uri= comp.task.restart

pub fn restart_frame_subtasks(
&self,
task_id: String,
frame: u32
) -> impl Future<Item = (), Error = Error>

RPC uri= comp.task.subtasks.frame.restart

pub fn restart_subtasks_from_task(
&self,
task_id: String,
subtask_ids: Vec<String>
) -> impl Future<Item = Value, Error = Error>

Restarts a set of subtasks from the given task. If the specified task is already finished, all failed subtasks will be restarted along with the set provided as a parameter. Finished subtasks will have their results copied over to the newly created task.

Parameters

task_id the ID of the task which contains the given subtasks.

subtask_ids the set of subtask IDs which should be restarted. If this is empty and the task is finished, all of the task's subtasks marked as failed will be restarted.

ignore_gas_price if True, this will ignore long transaction time errors and proceed with the restart.

disable_concent setting this flag to True will result in forcing Concent to be disabled for the task. This only has effect when the task is already finished and needs to be restarted.

##Returns

In case of any errors, returns the representation of the error (either a string or a dict). Otherwise, returns None.

RPC uri= comp.task.subtasks.restart

pub fn run_test_task(
&self,
task_spec: Value
) -> impl Future<Item = bool, Error = Error>

RPC uri= comp.tasks.check

pub fn check_test_status<'b>(
&'b self
) -> impl Future<Item = TaskTestResult, Error = Error> + 'static

RPC uri= comp.task.test.status

pub fn abort_test_task<'b>(
&'b self
) -> impl Future<Item = bool, Error = Error> + 'static

Returns true if there was task to cancel RPC uri= comp.tasks.check.abort

pub fn get_tasks_stats<'b>(
&'b self
) -> impl Future<Item = SubtaskStats, Error = Error> + 'static

RPC uri= comp.tasks.stats

pub fn get_environments<'b>(
&'b self
) -> impl Future<Item = Vec<CompEnvStatus>, Error = Error> + 'static

RPC uri= comp.environments

pub fn enable_environment(
&self,
env_id: String
) -> impl Future<Item = Option<String>, Error = Error>

Enables enviroment Returns None or Error message. RPC uri= comp.environment.enable

pub fn disable_environment(
&self,
env_id: String
) -> impl Future<Item = Option<String>, Error = Error>

Enables enviroment Returns None or Error message. RPC uri= comp.environment.disable

pub fn run_benchmark(
&self,
env_id: String
) -> impl Future<Item = Value, Error = Error>

RPC uri= comp.environment.benchmark

pub fn perf_mult_set(
&self,
multiplier: f64
) -> impl Future<Item = (), Error = Error>

RPC uri= performance.multiplier.update

pub fn perf_mult<'b>(
&'b self
) -> impl Future<Item = f64, Error = Error> + 'static

RPC uri= performance.multiplier

impl<'a, Inner: RpcEndpoint + ?Sized + 'static> GolemComp<'a, Inner>

pub fn create_task(
&self,
task_spec: Value
) -> impl Future<Item = String, Error = Error>